requirements.txt
pillow
streamlit
imgkit
python-magic